requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
click==8.3.1
distro==1.9.0
fastapi==0.128.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
openai==2.14.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
sniffio==1.3.1
starlette==0.50.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
